Further refinement on LINDetailActivity.  Fixed bugs with screen rotation.
Also implmented Parcable on Item, LIN and NSN.  Not sure if I'm going to need them, but the code is now there.  I'm starting to work away from long variables for the ids.  I don't remember why I did that in the first place, but it makes working with SparseArrays a pain in the ass.